feat: add button clear for H.9.2
