+++ src/components/card/_card-style.scss
@@ -1,15 +1,15 @@
 .card {
     margin: 0 auto;
     margin-bottom: 70px;
-    width: 600px;
+    max-width: 600px;
     min-height: 300px;
     padding: 20px 70px;
     border-radius: 55px;
     -webkit-box-shadow: -11px 10px 28px 0px rgba(0,0,0,0.45);
     -moz-box-shadow: -11px 10px 28px 0px rgba(0,0,0,0.45);
     box-shadow: -11px 10px 28px 0px rgba(0,0,0,0.45);
-    // backdrop-filter: blur(6px);
-    // -webkit-backdrop-filter: blur(6px);
+    backdrop-filter: blur(4px);
+    -webkit-backdrop-filter: blur(4px);
     // filter: grayscale(60%);
 
     //glass effect below:
