+++ src/components/task-H.9.3/TimePicker.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+class TimePicker extends React.Component {
+
+
+    render() {
+
+        console.log('render');
+
+        return (
+
+            console.log('return'),
+
+            <div>
+                hello
+            </div>
+        )
+    }
+}
+
+export default TimePicker
