+++ src/styles/scss/App.scss
@@ -28,4 +28,6 @@
 @import '../../components/task-T.1/checkList';
 @import '../../components/task-H.8/postRequest';
 @import '../../components/task-H.9.1/addRandomNames';
+@import '../../components/task-H.9.3/timePicker';
+@import '../../components/router/router';
